# docs/swagger.yaml
basePath: /
definitions:
...
  domain.User:
    properties:
      clicks_current_month:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      plan_expires:
        description: nil = forever (for free)
        type: string
      plan_id:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
    type: object
  domain.UserRole:
    enum:
    - admin
    - client
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleClient
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  v1.StatusData:
    properties:
      status:
        example: updated
        type: string
    type: object
  v1.StatusResponse:
    properties:
      data:
        $ref: '#/definitions/v1.StatusData'
    type: object
...
  v1.UpdateUserPlanReq:
    properties:
      plan_id:
        example: plan_pro
        type: string
    required:
    - plan_id
    type: object
  v1.UpdateUserStatusReq:
    properties:
      is_active:
        example: true
        type: boolean
    type: object
  v1.UsersListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.User'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Traffic analytics service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Traffic Analytics
  version: "1.0"
paths:
...
  /api/v1/admin/users:
    get:
      description: Get paginated users list (Admin). Returns array wrapped in data.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UsersListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Admin
  /api/v1/admin/users/{id}/plan:
    patch:
      consumes:
      - application/json
      description: Set new plan for user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Plan Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateUserPlanReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'Status: updated'
          schema:
            $ref: '#/definitions/v1.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change user plan
      tags:
      - Admin
  /api/v1/admin/users/{id}/status:
    patch:
      consumes:
      - application/json
      description: Activate/Deactivate user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Status Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateUserStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'Status: updated'
          schema:
            $ref: '#/definitions/v1.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user status
      tags:
      - Admin
...
securityDefinitions:
  APIKeyAuth:
    description: Static API Key for admin/service access
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"